server-flask/requirements.txt
Flask
flask-cors
numpy
python-dotenv
pyjwt
scikit-learn
gunicorn
